ref(tech): organize code with a library crate
Ref: #36
