+++ minigrep/src/lib.rs
@@ -1,0 +1,23 @@
+use std::{error::Error, fs};
+
+pub struct Config {
+    pub query: String,
+    pub file_path: String,
+}
+
+impl Config {
+    pub fn build(args: &[String]) -> Result<Self, &'static str> {
+        if args.len() < 3 {
+            return Err("not enough arguments");
+        }
+        let query = args[1].clone();
+        let file_path = args[2].clone();
+        Ok(Config { query, file_path })
+    }
+}
+
+pub fn run(config: Config) -> Result<(), Box<dyn Error>> {
+    let file_content = fs::read_to_string(config.file_path)?;
+    log::info!("With text:\n{file_content}");
+    Ok(())
+}
